test-requirements.txt
paho-mqtt==1.6.1
pytest
pytest-cov
pytest-asyncio
pytest-homeassistant-custom-component